remove ComponentStorage from detail namespace
